Task_7/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
pydantic==2.5.0
huggingface_hub==0.19.4
transformers==4.36.2